@@ -3,15 +3,27 @@ import 'package:get/get.dart';
 import '../controllers/my_home_page_controller.dart';
 import './about_page.dart';
 
-class MyHomePage extends StatelessWidget {
+class MyHomePage extends StatefulWidget {
   final String title;
+  MyHomePage({required this.title});
+
+  @override
+  _MyHomePageState createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
   final MyHomePageController controller = Get.put(MyHomePageController());
-  MyHomePage({this.title});
+  @override
+  void dispose() {
+    controller.dispose();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(title),
+        title: Text(widget.title),
       ),
       body: Center(
         child: Column(
@@ -26,26 +38,27 @@ class MyHomePage extends StatelessWidget {
                 style: Theme.of(context).textTheme.headline4,
               ),
             ),
-            FlatButton(
+            TextButton(
                 onPressed: () {
-                  Get.off(AboutPage());
+                  Get.to(() => AboutPage());
                 },
                 child: Text('About GetX')),
-            FlatButton(
+            TextButton(
                 onPressed: () {
                   Get.snackbar('GetX Snackbar', 'Yay! Awesome GetX Snackbar',
-                      snackPosition: SnackPosition.BOTTOM,
-                      backgroundColor: Colors.amberAccent);
+                      snackPosition: SnackPosition.BOTTOM, backgroundColor: Colors.amberAccent);
                 },
                 child: Text('Show Snackbar')),
-            FlatButton(
+            TextButton(
                 onPressed: () {
                   Get.defaultDialog(
-                      title: 'GetX Alert',
-                      middleText: 'Simple GetX alert',
-                      textConfirm: 'Okay',
-                      confirmTextColor: Colors.amberAccent,
-                      textCancel: 'Cancel');
+                    title: 'GetX Alert',
+                    middleText: 'Click Okay to reset counter',
+                    textConfirm: 'Okay',
+                    confirmTextColor: Colors.amberAccent,
+                    textCancel: 'Cancel',
+                    onConfirm: resetCounter,
+                  );
                 },
                 child: Text('Show AlertDialog'))
           ],
@@ -57,5 +70,10 @@ class MyHomePage extends StatelessWidget {
         child: Icon(Icons.add),
       ),
     );
+  }
+
+  void resetCounter() {
+    controller.count.value = 0;
+    Get.back();
   }
 }
